rename: room에 포함 되는 컴포넌트 하위 폴더로 경로 이동
