+++ src/components/room/meeting/place/Place.tsx
